# fssai-ra/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FSSAI-RA Control Plane",
    "description": "Reference API for exact-action proposals, human approval, execution, evidence, and recovery. Header identity is a replaceable teaching adapter.",
    "version": "0.5.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/profile": {
      "get": {
        "summary": "Get Profile",
        "operationId": "get_profile_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/resources": {
      "post": {
        "summary": "Register Resource",
        "operationId": "register_resource_v1_resources_post",
        "parameters": [
          {
            "name": "X-FSSAI-Identity",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Identity"
            }
          },
          {
            "name": "X-FSSAI-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/resources/{resource_id}": {
      "get": {
        "summary": "Get Resource",
        "operationId": "get_resource_v1_resources__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          },
          {
            "name": "X-FSSAI-Identity",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Identity"
            }
          },
          {
            "name": "X-FSSAI-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/proposals": {
      "post": {
        "summary": "Create Proposal",
        "operationId": "create_proposal_v1_proposals_post",
        "parameters": [
          {
            "name": "X-FSSAI-Identity",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Identity"
            }
          },
          {
            "name": "X-FSSAI-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/proposals/{request_id}/approval": {
      "post": {
        "summary": "Approve Proposal",
        "operationId": "approve_proposal_v1_proposals__request_id__approval_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "X-FSSAI-Identity",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Identity"
            }
          },
          {
            "name": "X-FSSAI-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/proposals/{request_id}/execute": {
      "post": {
        "summary": "Execute Proposal",
        "operationId": "execute_proposal_v1_proposals__request_id__execute_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "X-FSSAI-Identity",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Identity"
            }
          },
          {
            "name": "X-FSSAI-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/proposals/{request_id}/evidence": {
      "get": {
        "summary": "Get Evidence",
        "operationId": "get_evidence_v1_proposals__request_id__evidence_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "X-FSSAI-Identity",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Identity"
            }
          },
          {
            "name": "X-FSSAI-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reconcile": {
      "post": {
        "summary": "Reconcile",
        "operationId": "reconcile_v1_reconcile_post",
        "parameters": [
          {
            "name": "X-FSSAI-Identity",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Identity"
            }
          },
          {
            "name": "X-FSSAI-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Fssai-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApprovalCreate": {
        "properties": {
          "ttl_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Ttl Seconds",
            "default": 300
          }
        },
        "type": "object",
        "title": "ApprovalCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProposalCreate": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "operation": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Operation"
          },
          "resource_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Resource Id"
          },
          "from_status": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "From Status"
          },
          "to_status": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "To Status"
          },
          "evidence_version": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Evidence Version"
          }
        },
        "type": "object",
        "required": [
          "operation",
          "resource_id",
          "from_status",
          "to_status",
          "evidence_version"
        ],
        "title": "ProposalCreate"
      },
      "ResourceCreate": {
        "properties": {
          "resource_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Resource Id"
          },
          "status": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "resource_id",
          "status"
        ],
        "title": "ResourceCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}